Add gitignore

+++ efficientvit/samcore/trainer/medsam_trainer.py
@@ -187,7 +187,7 @@ class MedSAMTrainer(Trainer):
                 # pdb.set_trace()
                 print(output_i)
                 print(output_i.dtype)
-                loss_mask_i, loss_dice_i = loss_masks(output_i, masks, len(output_i), mode="none")
+                loss_mask_i, loss_dice_i = loss_masks(output_i, masks.to(torch.float32), len(output_i), mode="none")
                 loss_i = loss_mask_i * 20 + loss_dice_i
                 loss_list.append(loss_i)
             loss = torch.stack(loss_list, -1)
